Feature/gmail (#67)

+++ src/providers/BaseProvider.ts
@@ -2,7 +2,7 @@ import { explodeDID } from '@verida/helpers'
 import { Request, Response } from 'express'
 import { Utils } from '../utils'
 import serverconfig from '../serverconfig.json'
-import { AccountAuth, AccountProfile, BaseProviderConfig, Connection, SyncHandlerStatus, SyncProviderErrorEvent, SyncProviderLogEntry, SyncProviderLogLevel, SyncSchemaPosition, SyncSchemaPositionType, SyncStatus } from '../interfaces'
+import { AccountAuth, BaseProviderConfig, Connection, ConnectionProfile, SyncHandlerStatus, SyncProviderErrorEvent, SyncProviderLogEntry, SyncProviderLogLevel, SyncSchemaPosition, SyncSchemaPositionType, SyncStatus } from '../interfaces'
 import { IContext, IDatastore } from '@verida/types'
 import BaseSyncHandler from './BaseSyncHandler'
 import { SchemaRecord } from '../schemas'
@@ -13,11 +13,10 @@ const SCHEMA_SYNC_LOG = serverconfig.verida.schemas.SYNC_LOG
 export default class BaseProvider {
 
     protected config: BaseProviderConfig
-    protected vault?: IContext
+    protected vault: IContext
     protected connection?: Connection
     protected connectionDs?: IDatastore
     protected newAuth?: AccountAuth
-    protected profile?: AccountProfile
     
     public constructor(config: BaseProviderConfig, vault?: IContext, connection?: Connection) {
         this.config = config
@@ -64,46 +63,13 @@ export default class BaseProvider {
     public async callback(req: Request, res: Response, next: any): Promise<any> {
         throw new Error('Not implemented')
     }
-
-    public async getProfileData(did: string): Promise<Record<string, any>> {
-        const profileLabel = this.profile.name || this.profile.username || this.profile.id
-        const { address: didAddress } = explodeDID(did)
-
-        const credentialData: Record<string, any> = {
-            did,
-            didAddress: didAddress.toLowerCase(),
-            name: `${this.getProviderLabel()}: ${profileLabel}`,
-            type: `${this.getProviderId()}-account`,
-            image: this.getProviderSbtImage(),
-            description: `Proof of ${this.getProviderLabel()} account ownership ${profileLabel}${profileLabel == this.profile.id ? '' : ' (' + this.profile.id+ ')'}`,
-            attributes: [{
-                trait_type: "accountCreated",
-                value: this.profile.createdAt
-            }],
-            uniqueAttribute: this.profile.id,
-        }
-
-        if (this.profile.url) {
-            credentialData.external_url = this.profile.url
-        }
-
-        if (this.profile.avatarUrl) {
-            credentialData.attributes.push({
-                trait_type: "avatarUrl",
-                value: this.profile.avatarUrl
-            })
-        }
-
-        return credentialData
+    
+    public getProfile(): ConnectionProfile | undefined {
+        return this.connection.profile
     }
 
-    public async getProfile(did: string, context: IContext): Promise<AccountProfile> {
-        if (this.profile && !this.profile.credential) {
-            const profileCredentialData = await this.getProfileData(did)
-            this.profile.credential = await Utils.buildCredential(profileCredentialData, context)
-        }
-
-        return this.profile
+    public async getDatastore(schemaUrl: string): Promise<IDatastore> {
+        return this.vault.openDatastore(schemaUrl)
     }
 
     protected async logMessage(level: SyncProviderLogLevel, message: string, schemaUri?: string): Promise<void> {
@@ -123,7 +89,7 @@ export default class BaseProvider {
     /**
      * Reset this provider by deleting all position information and data
      */
-    public async reset(clearTokens: boolean = false): Promise<number> {
+    public async reset(deleteData: boolean = true, clearTokens: boolean = false): Promise<number> {
         const syncHandlers = await this.getSyncHandlers()
 
         let deletedRowCount = 0
@@ -147,19 +113,21 @@ export default class BaseProvider {
             }
             
             // delete data
-            const datastore = await this.vault.openDatastore(schemaUri)
-            while (true) {
-                const rows = <SchemaRecord[]> await datastore.getMany({
-                    sourceApplication: this.getProviderApplicationUrl()
-                })
+            if (deleteData) {
+                const datastore = await this.vault.openDatastore(schemaUri)
+                while (true) {
+                    const rows = <SchemaRecord[]> await datastore.getMany({
+                        sourceApplication: this.getProviderApplicationUrl()
+                    })
 
-                if (rows.length == 0) {
-                    break
-                }
+                    if (rows.length == 0) {
+                        break
+                    }
 
-                for (let r in rows) {
-                    await datastore.delete(rows[r]._id)
-                    deletedRowCount++
+                    for (let r in rows) {
+                        await datastore.delete(rows[r]._id)
+                        deletedRowCount++
+                    }
                 }
             }
         }
@@ -281,7 +249,7 @@ export default class BaseProvider {
     }
 
     public async getSyncHandler(handler: typeof BaseSyncHandler): Promise<BaseSyncHandler> {
-        return new handler(this.config, this.profile)
+        return new handler(this.config, this.connection, this)
     }
 
     /**
@@ -297,7 +265,7 @@ export default class BaseProvider {
         for (let h in handlers) {
             const handler = handlers[h]
             
-            const handlerInstance = new handler(this.config, this.profile)
+            const handlerInstance = new handler(this.config, this.connection, this)
             syncHandlers.push(handlerInstance)
         }
